fixed negative numbers in Drive-ESP32 API libraries

+++ Control_Subsystem/Arduino-ESP32 Interface/arduino_code/ControlInterface.cpp
@@ -32,7 +32,7 @@ int ControlInterface::fetchData()  {
         for(int i=0;i<ctrl_size;i++)    {
             high_byte=serial->read();
             low_byte=serial->read();
-            tmp[i]=(high_byte<<8)+low_byte;
+            tmp[i]=(int16_t)((high_byte<<8)+low_byte);
         }
         // load received control values into respective variables
         drive_mode=tmp[0];
@@ -66,7 +66,7 @@ int ControlInterface::getDriveMode() const  {
 }
 
 int ControlInterface::getDirection() const  {
-    return (int16_t)direction;
+    return direction;
 }
 
 int ControlInterface::getSpeed() const  {
@@ -74,15 +74,15 @@ int ControlInterface::getSpeed() const  {
 }
 
 int ControlInterface::getDistance() const   {
-    return (int16_t)distance;
+    return distance;
 }
 
 int ControlInterface::getTargetX() const    {
-    return (int16_t)target_x;
+    return target_x;
 }
 
 int ControlInterface::getTargetY() const    {
-    return (int16_t)target_y;
+    return target_y;
 }
 
 unsigned long ControlInterface::getSystemTime() const   {
